added activity_tspactivity.xml, DirectionsParser.java, TSP.java, TSPActivity.java
edited addplaces.java, AndroidManifest.xml, build.gradle app level, gradle.properties, strings.xml, IternaryActivity.java, MainActivity.java,
